requirements.txt

asgiref==3.6.0
cssselect==1.2.0
Django==4.2.1
django-jsignature==0.11
django-multiupload==0.6.1
gunicorn==20.1.0
lxml==4.9.2
Pillow==9.5.0
pyquery==2.0.0
sqlparse==0.4.4
mysqlclient==2.1.1


